# openapi.yml
openapi: 3.0.3
info:
  title: Pokedex V2 – Backend API
  version: 1.0.0
  description: |
    API REST para gestionar entrenadores, Pokémon y equipos.
    Incluye autenticación con JWT, gestión de equipos y relación Pokémon–Equipo.
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/pokemon:
    post:
      tags: [Pokemon]
      summary: Crear Pokémon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PokemonRequest'
            example:
              nombre: "Bulbasaur"
              tipo_principal: "Planta"
              tipo_secundario: "Veneno"
              sprite_normal: "https://img.poke/bulbasaur.png"
              sprite_shiny: "https://img.poke/bulbasaur_shiny.png"
      responses:
        '201':
          description: Pokémon creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
    get:
      tags: [Pokemon]
      summary: Obtener todos los Pokémon
      responses:
        '200':
          description: Lista de Pokémon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'

    patch:
      tags: [Pokemon]
      summary: Actualizar Pokémon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PokemonUpdateRequest'
      responses:
        '200':
          description: Pokémon actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.
  /api/v2/pokemon/nombre/{nombre}:
    get:
      tags: [Pokemon]
      summary: Obtener Pokémon por nombre
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pokémon encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.

    delete:
      tags: [Pokemon]
      summary: Eliminar Pokémon por nombre
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pokémon eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.
...
  /api/v2/pokemon/{id}:
    delete:
      tags: [Pokemon]
      summary: Eliminar Pokémon por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pokémon eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GeneralResponse:
      type: object
      properties:
        uri:
          type: string
          nullable: true
        message:
          type: string
        status:
          type: integer
          format: int32
        date:
          type: string
          format: date
        data:
          description: Contenido variable de la respuesta (puede ser objeto, lista o valor simple).
      required: [message, status]
...
    PokemonRequest:
      type: object
      properties:
        nombre:
          type: string
        tipo_principal:
          type: string
        tipo_secundario:
          type: string
          nullable: true
        sprite_normal:
          type: string
          nullable: true
        sprite_shiny:
          type: string
          nullable: true
      required:
        - nombre
        - tipo_principal
    PokemonUpdateRequest:
      type: object
      properties:
        id_pokemon:
          type: integer
          format: int64
        tipo_principal:
          type: string
        tipo_secundario:
          type: string
          nullable: true
        sprite_normal:
          type: string
          nullable: true
        sprite_shiny:
          type: string
          nullable: true
      required:
        - id_pokemon
...